requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
openai==1.82.0
httpx==0.28.1
chromadb==1.0.7
sentence-transformers==4.1.0
pydantic==2.11.3
aiosqlite==0.21.0
slowapi==0.1.9
